Revert jukebox dispenser behavior
Moved this to a plugin, which makes it unnecessary to have this patch in the server software.

Additionally, it was not possible to add a configuration toggle for this feature, since the behavior is statically loaded before the configuration file loads in.

Having this in a plugin makes more sense.